+++ CHANGELOG.rst
@@ -16,6 +16,7 @@ Latest versions
 
 Version 1.4.1 (2020-05-29)
 --------------------------
+- Added ``sim.people.plot()``, which shows the age distribution, and distribution of contacts by age and layer.
 - Added ``sim.make_age_histogram()``, as well as the ability to call ``cv.age_histogram(sim)``, as an alternative to adding these as analyzers to a sim.